[WIP] add courses list view (Fixes #83)
