+++ src/app/courses/add/constants.ts
@@ -1,0 +1,42 @@
+export const gradeLevels = [
+  'Pre-Kindergarten',
+  'Kindergarten',
+  '1',
+  '2',
+  '3',
+  '4',
+  '5',
+  '6',
+  '7',
+  '8',
+  '9',
+  '10',
+  '11',
+  '12',
+  'College',
+  'Post-Graduate'
+];
+
+export const subjectLevels = ['Beginner', 'Intermediate', 'Advanced', 'Expert'];
+
+export const days = [
+  'Saturday',
+  'Sunday',
+  'Monday',
+  'Tuesday',
+  'Wednesday',
+  'Thursday',
+  'Friday'
+];
+
+export default function searchDocuments(selector, query) {
+  return JSON.parse(`
+    {
+      "selector": {
+        "${selector}": "${query}"
+      },
+      "fields": ["${selector}"],
+      "limit": 1
+    }
+    `);
+}
