+++ src/app/courses/add/coursesadd.component.scss
@@ -1,0 +1,19 @@
+@import '../../variables';
+
+/* from the angular docs */
+// moved it here for now because it may cause issues with other forms
+.ng-valid[required],
+.ng-dirty.ng-valid:not(form) {
+  border-left: 5px solid $ole-green;
+  border-color: $ole-green;
+}
+
+.ng-dirty.ng-invalid:not(form) {
+  border-left: 5px solid $error-red;
+  border-color: $error-red;
+}
+
+.ng-dirty.ng-pending:not(form) {
+  border-left: 5px solid $pending;
+  border-color: $pending;
+}
