+++ src/app/courses/courses.component.ts
@@ -1,147 +1,34 @@
-import { Component, OnDestroy } from '@angular/core';
-import {
-  FormBuilder,
-  FormControl,
-  FormGroup,
-  FormArray,
-  Validators
-} from '@angular/forms';
-import { Location } from '@angular/common';
-import { Router } from '@angular/router';
-
+import { Component, OnInit } from '@angular/core';
 import { CouchService } from '../shared/couchdb.service';
-import { CustomValidators } from '../validators/custom-validators';
-import { CourseValidatorService } from '../validators/course-validator.service';
-// searchDocuments is declared as a default export so we can import it like this
-import searchDocuments, * as constants from './constants';
 
 @Component({
   selector: 'planet-courses',
   templateUrl: './courses.component.html',
   styleUrls: ['./courses.component.scss']
 })
-export class CoursesComponent {
-  // needs member document to implement
-  members = [];
-  readonly dbName = 'courses'; // make database name a constant
-
-  courseForm: FormGroup;
-
-  showDaysCheckBox = true; // for toggling the days checkbox
-
-  // from the constants import
-  gradeLevels = constants.gradeLevels;
-  subjectLevels = constants.subjectLevels;
-  days = constants.days;
-
+export class CoursesComponent implements OnInit {
+  message = '';
+  courses = [];
   constructor(
-    private location: Location,
-    private router: Router,
-    private fb: FormBuilder,
-    private couchService: CouchService,
-    private courseValidatorService: CourseValidatorService
-  ) {
-    this.createForm();
+      private couchService: CouchService
+  ) { }
+
+  getCourses() {
+    this.couchService.get('courses/_all_docs?include_docs=true')
+        .then((data) => {
+          // don't retrieve the course validator
+          this.courses = data.rows.filter(x => x.doc._id !== '_design/course-validators');
+        }, (error) => this.message = 'There was a problem getting the courses');
   }
 
-  createForm() {
-    this.courseForm = this.fb.group({
-      courseTitle: [
-        '',
-        Validators.required,
-        // an arrow function is for lexically binding 'this' otherwise 'this' would be undefined
-        ac => this.courseValidatorService.checkCourseExists$(ac)
-      ],
-      description: ['', Validators.required],
-      languageOfInstruction: '',
-      memberLimit: [
-        '', // need to compose validators if we use more than one
-        Validators.compose([
-          CustomValidators.integerValidator,
-          Validators.min(1)
-        ])
-      ],
-      courseLeader: [''],
-      method: '',
-      gradeLevel: '',
-      subjectLevel: '',
-      startDate: ['', CustomValidators.dateValidator],
-      endDate: [
-        '',
-        Validators.compose([
-          // we are using a higher order function so we  need to call the validator function
-          CustomValidators.endDateValidator(),
-          CustomValidators.dateValidator
-        ])
-      ],
-      day: this.fb.array([]),
-      startTime: ['', CustomValidators.timeValidator],
-      endTime: [
-        '',
-        Validators.compose([
-          CustomValidators.endTimeValidator(),
-          CustomValidators.timeValidator
-        ])
-      ],
-      location: '',
-      backgroundColor: ['', CustomValidators.hexValidator],
-      foregroundColor: ['', CustomValidators.hexValidator]
-    });
-
-    // set default values
-    this.courseForm.patchValue({
-      gradeLevel: this.gradeLevels[0],
-      subjectLevel: this.subjectLevels[0],
-      memberLimit: 10,
-      backgroundColor: '#ffffff',
-      foregroundColor: '#000000'
-    });
+  deleteCourse(courseId, courseRev) {
+    this.couchService.delete('courses/' + courseId + '?rev=' + courseRev)
+        .then((data) => {
+          this.getCourses();
+        }, (error) => this.message = 'There was a problem deleting this course');
   }
 
-  onSubmit() {
-    if (this.courseForm.valid) {
-      this.addCourse(this.courseForm.value);
-    } else {
-      Object.keys(this.courseForm.controls).forEach(field => {
-        const control = this.courseForm.get(field);
-        control.markAsTouched({ onlySelf: true });
-      });
-    }
-  }
-
-  async addCourse(courseInfo) {
-    // ...is the rest syntax for object destructuring
-    await this.couchService.post(this.dbName, { ...courseInfo });
-    this.router.navigate(['/']);
-  }
-
-  cancel() {
-    this.location.back();
-  }
-
-  /* FOR TOGGLING DAILY/WEEKLY DAYS */
-
-  onDayChange(day: string, isChecked: boolean) {
-    const dayFormArray = <FormArray>this.courseForm.controls.day;
-
-    if (isChecked) {
-      // add to day array if checked
-      dayFormArray.push(new FormControl(day));
-    } else {
-      // remove from day array if unchecked
-      const index = dayFormArray.controls.findIndex(x => x.value === day);
-      dayFormArray.removeAt(index);
-    }
-  }
-
-  // remove old values from array on radio button change
-  toogleWeekly(val: boolean) {
-    // empty the array
-    this.courseForm.setControl('day', this.fb.array([]));
-    if (val) {
-      // add all days to the array if the course is daily
-      this.courseForm.setControl('day', this.fb.array(this.days));
-    }
-    this.showDaysCheckBox = val;
+  ngOnInit() {
+    this.getCourses();
   }
 }
