+++ src/app/courses/request/coursesrequest.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'planet-courses',
+  templateUrl: 'coursesrequest.component.html',
+  styleUrls: ['coursesrequest.component.scss']
+})
+export class CoursesrequestComponent {}
